pyproject.toml
[project]
name = "llm-tournament"
version = "0.2.0"
description = "LLM Tournament Evaluator - Compare OpenRouter models via pairwise Elo/TrueSkill ranking"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "you@example.com" }]
dependencies = [
    "typer>=0.9.0",
    "httpx>=0.25.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "structlog>=23.2.0",
    "tenacity>=8.2.0",
    "rich>=13.7.0",
    "anyio>=4.0.0",
    "duckdb>=1.0.0",
    "openskill>=5.0.0",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "sqlmodel>=0.0.31",
    "duckdb-engine>=0.17.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[project.scripts]
llm-tournament = "llm_tournament.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_tournament"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "C4", "DTZ", "T10", "ISC", "ICN", "PIE", "PT", "RET", "SIM", "TID", "ARG", "PL", "TRY", "RUF"]
ignore = ["PLR0913", "TRY003", "S101"]
